feat: make P&L story decorator configurable and drop container for summaries
